* performance improvements:   * make a stencil pass prior to using the light pass shader to determine which pixels must be shaded by a light sphere (non camera intersecting/enclosing lights only)   * depth testing optimizations (+some reordering), try to use GL_LESS everywhere and try to reuse the geometry buffer depth buffer in later passes (light + material pass)  * added a gl_timer class (a wrapper around GL_ARB_query_timer) and use it throughout the engine  * added actual stencil attachment/buffer support to the rtt class (only packed depth+stencil is allowed by all? graphics cards)  * reintegrated wip opencl light pass (disabled for now), note that gl depth buffer sharing is not spec conformant (-> needs to make a copy every time)  * fixed minimum particle requirement (this can be any number now and is not restricted by the local wg size; note: pot particle count rounding still applies)  * added a new and simpler light sphere (blender icosphere with 2x subdivision; note: this must be scaled by 1.07 to encompass a sphere of radius 1)  * fixed scene and opencl construction/destruction order (scene depends on opencl)  * switched to consistent GL_LESS depth testing when rendering 3d (2d should still use GL_LEQUAL)  * updated opencl error codes (for opencl 1.2)  * fixed opencl buffer deletion  * for extra precision: use doubles in camera class to detect and handle x/y mouse rotation/movement  * fixed possible rounding error (off by one) when computing the final buffer size in scene::recreate_buffers

+++ src/scene/camera.cpp
@@ -68,12 +68,12 @@ void camera::run() {
 #if !defined(__APPLE__)
 		SDL_GetMouseState(&cursor_pos_x, &cursor_pos_y);
 
-		float xpos = (1.0f / (float)e->get_width()) * (float)cursor_pos_x;
-		float ypos = (1.0f / (float)e->get_height()) * (float)cursor_pos_y;
+		double xpos = (1.0 / (double)e->get_width()) * (double)cursor_pos_x;
+		double ypos = (1.0 / (double)e->get_height()) * (double)cursor_pos_y;
 		
-		if(xpos != 0.5f || ypos != 0.5f) {
-			rotation.x -= (0.5f - ypos) * rotation_speed;
-			rotation.y -= (0.5f - xpos) * rotation_speed;
+		if(xpos != 0.5 || ypos != 0.5) {
+			rotation.x -= (0.5 - ypos) * (double)rotation_speed;
+			rotation.y -= (0.5 - xpos) * (double)rotation_speed;
 			SDL_WarpMouseInWindow(e->get_window(), e->get_width()/2, e->get_height()/2);
 		}
 ////////////////////////////////
@@ -81,13 +81,13 @@ void camera::run() {
 #else
 		SDL_GetRelativeMouseState(&cursor_pos_x, &cursor_pos_y);
 		
-		float xpos = (1.0f / (float)e->get_width()) * (float)-cursor_pos_x;
-		float ypos = (1.0f / (float)e->get_height()) * (float)-cursor_pos_y;
+		double xpos = (1.0 / (double)e->get_width()) * (double)-cursor_pos_x;
+		double ypos = (1.0 / (double)e->get_height()) * (double)-cursor_pos_y;
 		
-		if(xpos != 0.0f || ypos != 0.0f) {
+		if(xpos != 0.0 || ypos != 0.0) {
 			if(!ignore_next_rotation) {
-				rotation.x -= ypos * rotation_speed;
-				rotation.y -= xpos * rotation_speed;
+				rotation.x -= ypos * (double)rotation_speed;
+				rotation.y -= xpos * (double)rotation_speed;
 			}
 			else ignore_next_rotation--;
